Cleaning up warnings for C++11 standard

+++ Source/DREAM3DLib/DataArrays/StringDataArray.hpp
@@ -93,7 +93,7 @@ class StringDataArray : public IDataArray
     virtual IDataArray::Pointer deepCopy()
     {
       StringDataArray::Pointer daCopy = StringDataArray::CreateArray(GetNumberOfTuples(), GetName());
-      for(int32_t i = 0; i < m_Array.size(); ++i)
+      for(size_t i = 0; i < m_Array.size(); ++i)
       {
         daCopy->SetValue(i, m_Array[i]);
       }
